+++ apps/dv-admin/src/views/search/address/components/addressInfo/types.ts
@@ -1,8 +1,11 @@
 import type { ITransactionDashboardItem, IWalletInfoResponse } from "@dv-admin/utils/types/api/apiGo";
+import type { UItableMeta } from "@dv.net/ui-kit/dist/components/UiTable/types";
 
 export interface IProps {
-	walletInfo: IWalletInfoResponse | null;
-	transactionsSearchAddress: ITransactionDashboardItem[];
-	transferHistory: ITransactionDashboardItem[];
-	isLoadingTransactions: boolean;
+	walletInfo: IWalletInfoResponse | null
+	transactionsSearchAddress: ITransactionDashboardItem[]
+	transactionsSearchAddressPagination: UItableMeta | null
+	transferHistory: ITransactionDashboardItem[]
+	transferHistoryPagination: UItableMeta | null
+	isLoadingTransactions: boolean
 }